// client/src/global.types.ts
export interface ISignupParams {
	username : string;
	email: string;
	password: string;
	deposit?: number;
}